refactor: remove dead code, unused deps, and stale exports
Dead-code sweep guided by knip, now kept clean via a checked-in knip.json:

- Delete orphaned exchange-rate helpers (getExchangeRate,
  resolveMissingRates, persistExchangeRate) left behind by the
  resolveRate/cross-rate refactor (#396/#397)
- Drop the getNetWorthSummary backward-compat alias; snapshot-service
  calls getCachedNetWorthSummary directly
- Delete disabled prisma/seed.ts stub and four unreferenced shadcn
  primitives (checkbox, currency-cell, popover, separator)
- Delete unused chart tooltip/legend formatter factories and rename
  chart-formatters.tsx -> .ts (no JSX remains)
- Delete unused bouncySpringConfig, isOccSymbol, CurrencyCode,
  AllocationItem; un-export internal-only helpers
- Remove unused deps (@prisma/pg-worker, @prisma/ppg, @parcel/watcher);
  add previously-unlisted @radix-ui/react-slot, dotenv, @next/env
- Fix stale CLAUDE.md references (net-worth read name, rate pipeline
  description, chart-formatters path)
