+++ knip.json
@@ -1,0 +1,10 @@
+{
+  "$schema": "https://unpkg.com/knip@6/schema.json",
+  "ignore": [
+    "src/components/ui/**",
+    "tests/e2e/global-setup.ts",
+    "tests/e2e/global-teardown.ts",
+    "scripts/generate-splash-screens.mjs"
+  ],
+  "ignoreBinaries": ["docker-compose"]
+}
