+++ package.json
@@ -41,12 +41,10 @@
     "@base-ui/react": "^1.3.0",
     "@neondatabase/serverless": "^1.0.2",
     "@next/bundle-analyzer": "^16.2.4",
-    "@parcel/watcher": "^2.5.6",
     "@prisma/adapter-neon": "^7.6.0",
     "@prisma/adapter-pg": "^7.6.0",
     "@prisma/client": "^7.6.0",
-    "@prisma/pg-worker": "^6.9.0",
-    "@prisma/ppg": "^1.0.1",
+    "@radix-ui/react-slot": "^1.2.5",
     "@vercel/analytics": "^2.0.1",
     "@vercel/speed-insights": "^2.0.0",
     "class-variance-authority": "^0.7.1",
@@ -77,12 +75,14 @@
     "zod": "^4.3.6"
   },
   "devDependencies": {
+    "@next/env": "^16.2.9",
     "@playwright/test": "^1.52.0",
     "@tailwindcss/postcss": "^4",
     "@types/node": "^20",
     "@types/react": "^19",
     "@types/react-dom": "^19",
     "@types/ws": "^8.18.1",
+    "dotenv": "^17.4.2",
     "eslint-config-prettier": "^10.1.8",
     "husky": "^9.1.7",
     "lint-staged": "^17.0.3",
